+++ examples/FLORIS_power_comp/power_comp.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 
 import openmdao.api as om
 
-import ard.utils.core
+import ard.utils.io
 import ard.wind_query as wq
 import ard.farm_aero.floris as farmaero_floris
 
@@ -31,7 +31,7 @@ filename_turbine_spec = Path(
     "data",
     "turbine_spec_IEA-3p4-130-RWT.yaml",
 )  # toolset generalized turbine specification
-data_turbine_spec = ard.utils.core.load_turbine_spec(filename_turbine_spec)
+data_turbine_spec = ard.utils.io.load_turbine_spec(filename_turbine_spec)
 
 # set up the modeling options
 modeling_options = {
