+++ test/system/ard/collection/test_optiwindnet.py
@@ -14,7 +14,7 @@ from wisdem.optimization_drivers.nlopt_driver import NLoptDriver
 
 import ard
 import ard.utils.test_utils
-import ard.utils.core
+import ard.utils.io
 import ard.wind_query as wq
 import ard.layout.sunflower as sunflower
 import ard.farm_aero.floris as farmaero_floris
@@ -46,7 +46,7 @@ class TestoptiwindnetLayout:
             / "data"
             / "turbine_spec_IEA-3p4-130-RWT.yaml"
         )  # toolset generalized turbine specification
-        data_turbine_spec = ard.utils.core.load_turbine_spec(filename_turbine_spec)
+        data_turbine_spec = ard.utils.io.load_turbine_spec(filename_turbine_spec)
 
         # set up the modeling options
         self.modeling_options = {
