+++ test/unit/ard/layout/test_spacing.py
@@ -3,7 +3,7 @@ import numpy as np
 import jax
 import ard.layout.spacing
 import jax.numpy as jnp
-import ard.utils.core as utils
+import ard.utils.io as utils
 import openmdao.api as om
