+++ test/unit/utils/test_utils.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 import jax
 import jax.numpy as jnp
-import ard.utils.core as utils
+import ard.utils.io as utils
 
 
 class TestUtils:
